backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
sqlalchemy
aiomysql
pydantic[email]
python-jose[cryptography]
bcrypt
python-multipart
pymysql
